Coching-Cosmetic-AI-V1/demo-v1/requirements.txt
langchain>=0.1.0,<0.2.0
langchain-openai>=0.0.5,<0.1.0
langchain-core>=0.1.0,<0.2.0
langchain-text-splitters>=0.0.1,<0.1.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
openai>=1.0.0,<2.0.0
requests>=2.25.0,<3.0.0
tiktoken>=0.5.0,<1.0.0
elasticsearch>=8.0.0,<9.0.0
fastapi>=0.100.0,<1.0.0
uvicorn>=0.20.0,<1.0.0
